requirements.txt
requests~=2.31.0
google-cloud-bigquery==3.13.0
google-cloud-storage==2.13.0
pandas==2.1.4
pyarrow==13.0.0